fix conversion warning
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@1023 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dope/dope_main.ss/src/dope_bin2xml/dope_bin2xml.cpp
@@ -340,7 +340,7 @@ void ConvertFiles()
             while (file.good())
             {
                 file.read(&bin[0] + numBytesRead,4096);
-                numBytesRead += file.gcount();
+                numBytesRead += static_cast<size_t>(file.gcount());
             }
 
             if(fileSize != numBytesRead)
